Mistakenly Cloned my Repositories Before Forking
Originally, I made the very idiotic mistake of missing a large portion of the pre-course work by only doing the repositories (such as underbar, koans, ect) and not seeing the lessons.
Because of this I did not see that I was supposed to fork my repos first and the nclone them, instead I just clone them. This is what I'm doing now, and I'm essentially transferring my old repositories to the correct forked ones